requirements.txt
pymongo==4.5.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pyyaml==6.0.1
matplotlib==3.7.2
numpy==1.24.3
pybind11>=2.10.0
colorama==0.4.6